// src/modules/giphy/interfaces/giphy.dto.ts
export interface GiphyDto {
  analytics: unknown;
  analytics_response_payload: string;
  bitly_gif_url: string;
  bitly_url: string;
  content_url: string;
  embed_url: string;
  id: string;
  images: unknown;
  import_datetime: string;
  is_sticker: boolean;
  rating: string;
  slug: string;
  source: string;
  source_post_url: string;
  source_tld: string;
  title: string;
  trending_datetime: string;
  type: string;
  url: string;
  user: unknown;
  username: string;
}